refact(api): resolve schema name collisions from #5799 domain split (#5935 #5936 #5937)
- #5935: move AgentMessageResponse from schemas_agent to schemas_common
  (used by both agent.py and logs.py — cross-domain)
- #5936: remove dead analytics_cost.UsageRecordResponse; add
  CostTrackingRecordResponse to schemas_analytics with correct fields
- #5937: rename intelligent_agent.SystemInfoResponse →
  AgentSystemCapabilitiesResponse; move to schemas_agent; update 3 usages
