+++ autobot-backend/api/logs.py
@@ -21,7 +21,7 @@ import aiofiles
 from fastapi import APIRouter, Depends, HTTPException, Query, WebSocket
 from fastapi.responses import StreamingResponse
 
-from api.schemas_agent import AgentMessageResponse
+from api.schemas_common import AgentMessageResponse
 from api.schemas_code import (
     LogContainerResponse,
     LogReadResponse,
